requirements.txt
aiohttp==3.8.3
python-telegram-bot==20.0
fpdf==1.7.2
nest-asyncio==1.5.6
python-dotenv==1.0.0
asyncio==3.4.3
